+++ server/odc-service/src/main/java/com/oceanbase/odc/service/flow/task/model/DBStructureComparisonResp.java
@@ -17,10 +17,6 @@ package com.oceanbase.odc.service.flow.task.model;
 
 import java.util.List;
 
-import com.oceanbase.odc.core.flow.model.FlowTaskResult;
-import com.oceanbase.odc.core.shared.constant.TaskStatus;
-import com.oceanbase.odc.metadb.task.TaskEntity;
-import com.oceanbase.odc.service.structurecompare.DBObjectStructureComparisonResult;
 import com.oceanbase.tools.dbbrowser.model.DBObjectType;
 
 import lombok.Data;
@@ -31,20 +27,27 @@ import lombok.Data;
  * @since ODC_release_4.2.4
  */
 @Data
-public class StructureComparisonTaskResult implements FlowTaskResult {
-    /**
-     * Refer to {@link TaskEntity#getId()}
-     */
+public class DBStructureComparisonResp {
     private Long taskId;
-    private TaskStatus status;
-    private List<Comparing> comparingList;
+    private List<ComparisonResult> comparisonResults;
+    private String totalChangeScript;
+    private String storageObjectId;
 
-    private class Comparing {
+    private class ComparisonResult {
         /**
-         * Refer to {@link DBObjectStructureComparisonResult#getId()}
+         * Refer to {@link DBObjectStructureComparisonResp#getId()}
          */
         private Long structureComparisonId;
         private DBObjectType dbObjectType;
         private String dbObjectName;
+        private OperationType operationType;
+    }
+    public enum OperationType {
+        CREATE,
+        UPDATE,
+        DROP,
+        NO_ACTION,
+        SKIP,
+        UNSUPPORTED
     }
 }
